+++ RealmServer/PacketReader/CMSG_ITEM_QUERY_SINGLE.cs
@@ -7,16 +7,17 @@ namespace RealmServer.PacketReader
     /// </summary>
     public sealed class CMSG_ITEM_QUERY_SINGLE : Common.Network.PacketReader
     {
+        public uint ItemEntry;
         public ulong ItemId;
-        public uint ItemIdo;
+        
 
         public CMSG_ITEM_QUERY_SINGLE(byte[] data) : base(data)
         {
-            ItemIdo = ReadUInt32();
+            ItemEntry = ReadUInt32();
             ItemId = ReadUInt64();
 
 #if DEBUG
-            Log.Print(LogType.Debug, $"[CMSG_ITEM_QUERY_SINGLE] ItemIdo: {ItemIdo} ItemId: {ItemId}");
+            Log.Print(LogType.Debug, $"[CMSG_ITEM_QUERY_SINGLE] ItemEntry: {ItemEntry} ItemId: {ItemId}");
 #endif
         }
     }